transform the data to TableCard rows
